refactor(web-ui): replace nav shell with global search and session list
- Extract GlobalSearchDialog, SessionList, WorkspaceFooterActions from NavPanel/NavBar

- Wire UnifiedTopBar, WorkspaceBody, and scenes to new layout

- Remove ImageAnalysisCard; adjust ide-control and notification styles
