+++ src/web-ui/src/app/components/GlobalSearchDialog/GlobalSearchDialog.scss
@@ -14,8 +14,6 @@
   padding-top: 80px;
 }
 
-// ── Card — matches .bitfun-scene-viewport surface (--color-bg-scene)
-
 .bitfun-nav-search-dialog__card {
   width: 520px;
   max-width: calc(100vw - 32px);
@@ -53,13 +51,12 @@
     opacity: 0;
     transform: translateY(-12px) scale(0.982);
   }
+
   to {
     opacity: 1;
     transform: translateY(0) scale(1);
   }
 }
-
-// ── Input row ────────────────────────────────────────────────────────────────
 
 .bitfun-nav-search-dialog__input-row {
   padding: $size-gap-2;
@@ -68,7 +65,6 @@
   background: var(--color-bg-scene);
 }
 
-// Real search field in dialog (capsule). Nav trigger is a plain button — styles differ on purpose.
 .bitfun-nav-search-dialog__search {
   width: 100%;
 
@@ -114,13 +110,12 @@
     opacity: 0;
     transform: translateY(5px);
   }
+
   to {
     opacity: 1;
     transform: translateY(0);
   }
 }
-
-// ── Results ──────────────────────────────────────────────────────────────────
 
 .bitfun-nav-search-dialog__results {
   overflow-y: auto;
@@ -148,8 +143,6 @@
   font-size: var(--font-size-sm);
   animation: bitfun-nav-search-fade-rise 0.32s cubic-bezier(0.22, 1, 0.36, 1) 0.06s both;
 }
-
-// ── Groups ───────────────────────────────────────────────────────────────────
 
 .bitfun-nav-search-dialog__group {
   animation: bitfun-nav-search-fade-rise 0.34s cubic-bezier(0.22, 1, 0.36, 1) both;
@@ -184,7 +177,6 @@
   opacity: 0.88;
 }
 
-// Default (no query): sessions column — label only, no rows; hint below label.
 .bitfun-nav-search-dialog__session-hint {
   padding: $size-gap-2 $size-gap-2 $size-gap-1 2px;
   font-size: var(--font-size-xs);
@@ -192,8 +184,6 @@
   color: var(--color-text-muted);
   user-select: none;
 }
-
-// ── Result item ──────────────────────────────────────────────────────────────
 
 .bitfun-nav-search-dialog__item {
   display: flex;
