--- src/web-ui/src/app/components/NavBar/NavBar.scss
@@ -1,107 +1,0 @@
-/**
- * NavBar styles — back/forward + drag region + window controls (32px).
- */
-
-@use '../../../component-library/styles/tokens.scss' as *;
-
-.bitfun-nav-bar {
-  display: flex;
-  align-items: center;
-  height: 40px;
-  flex-shrink: 0;
-  padding: 0 $size-gap-2;
-  gap: $size-gap-1;
-  background: var(--color-bg-primary);
-  user-select: none;
-
-  &--collapsed {
-    justify-content: flex-start;
-    // Add tiny horizontal breathing room so edge buttons keep full corner rendering.
-    padding: 0 calc(#{$size-gap-2} + 1px);
-  }
-
-  &--macos {
-    // Reserve left native titlebar area for traffic lights in overlay mode.
-    padding-left: calc(#{$size-gap-2} + 72px);
-    padding-right: 2px;
-
-  }
-
-  &--macos#{&}--collapsed {
-    padding-left: calc(#{$size-gap-2} + 1px);
-    padding-right: 2px;
-  }
-
-  // ── Back / Forward buttons ──────────────────────────────
-
-  &__btn {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 26px;
-    height: 26px;
-    padding: 0;
-    border: none;
-    border-radius: 4px;
-    background: transparent;
-    color: var(--color-text-muted);
-    cursor: pointer;
-    flex-shrink: 0;
-    transition: color $motion-fast $easing-standard,
-                background $motion-fast $easing-standard;
-
-    &:hover:not(.is-inactive) {
-      color: var(--color-text-primary);
-      background: var(--element-bg-soft);
-    }
-
-    &.is-inactive {
-      opacity: 0.3;
-      cursor: default;
-    }
-
-    &:focus-visible {
-      outline: 2px solid var(--color-accent-500);
-      outline-offset: -1px;
-    }
-  }
-
-  &__panel-toggle {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    width: 26px;
-    height: 26px;
-    padding: 0;
-    border: none;
-    border-radius: 4px;
-    background: transparent;
-    color: var(--color-text-muted);
-    cursor: pointer;
-    transition: color $motion-fast $easing-standard,
-                background $motion-fast $easing-standard;
-
-    &:hover {
-      color: var(--color-text-primary);
-      background: var(--element-bg-soft);
-    }
-
-    &:focus-visible {
-      outline: 2px solid var(--color-accent-500);
-      outline-offset: -1px;
-    }
-  }
-
-  // ── WindowControls container ────────────────────────────
-  // WindowControls renders its own wrapper; just align it here.
-  .window-controls {
-    flex-shrink: 0;
-  }
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .bitfun-nav-bar__btn,
-  .bitfun-nav-bar__panel-toggle {
-    transition: none;
-  }
-}
