--- src/web-ui/src/app/components/NavBar/index.ts
@@ -1,1 +1,0 @@
-export { default as NavBar } from './NavBar';
