--- src/web-ui/src/app/components/NavPanel/index.ts
@@ -1,2 +1,0 @@
-export { default } from './NavPanel';
-export type { NavItem, NavSection, InlineSectionProps } from './types';
