--- src/web-ui/src/app/components/NavPanel/sections/sessions/SessionsSection.tsx
@@ -1,563 +1,0 @@
-/**
- * SessionsSection — inline accordion content for the "Sessions" nav item.
- *
- * Rendered inside NavPanel when the Sessions item is expanded.
- * Owns all data fetching / mutation for chat sessions.
- */
-
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { createPortal } from 'react-dom';
-import { Pencil, Trash2, Check, X, Code2, ListTodo, Sparkles, MoreHorizontal, Loader2 } from 'lucide-react';
-import { IconButton, Input, Tooltip } from '@/component-library';
-import { useI18n } from '@/infrastructure/i18n';
-import { flowChatStore } from '../../../../../flow_chat/store/FlowChatStore';
-import { flowChatManager } from '../../../../../flow_chat/services/FlowChatManager';
-import type { FlowChatState, Session } from '../../../../../flow_chat/types/flow-chat';
-import { useOverlayStore } from '../../../../stores/overlayStore';
-import { getWorkspaceDisplayName, useWorkspaceContext } from '@/infrastructure/contexts/WorkspaceContext';
-import { createLogger } from '@/shared/utils/logger';
-import { useAgentCanvasStore } from '@/app/components/panels/content-canvas/stores';
-import {
-  openBtwSessionInAuxPane,
-  openMainSession,
-  selectActiveBtwSessionTab,
-} from '@/flow_chat/services/openBtwSession';
-import { resolveSessionRelationship } from '@/flow_chat/utils/sessionMetadata';
-import {
-  compareSessionsForDisplay,
-  findOpenedWorkspaceForSession,
-  sessionBelongsToWorkspaceNavRow,
-} from '@/flow_chat/utils/sessionOrdering';
-import { stateMachineManager } from '@/flow_chat/state-machine';
-import { SessionExecutionState } from '@/flow_chat/state-machine/types';
-import './SessionsSection.scss';
-
-const log = createLogger('SessionsSection');
-const AGENT_SCENE = 'session' as const;
-
-type SessionMode = 'code' | 'cowork' | 'claw';
-
-const resolveSessionModeType = (session: Session): SessionMode => {
-  const normalizedMode = session.mode?.toLowerCase();
-  if (normalizedMode === 'cowork') return 'cowork';
-  if (normalizedMode === 'claw') return 'claw';
-  return 'code';
-};
-
-const getTitle = (session: Session): string =>
-  session.title?.trim() || `Task ${session.sessionId.slice(0, 6)}`;
-
-interface SessionsSectionProps {
-  workspaceId?: string;
-  workspacePath?: string;
-  /** Remote SSH: same `workspacePath` on different hosts must filter by this (see Session.remoteConnectionId). */
-  remoteConnectionId?: string | null;
-  /** Remote SSH: disambiguates same path on different hosts; when set with matching session host, connectionId may differ. */
-  remoteSshHost?: string | null;
-  isActiveWorkspace?: boolean;
-  showCreateActions?: boolean;
-  /** When set (e.g. assistant workspace), session row tooltip includes this assistant name. */
-  assistantLabel?: string;
-  /** When false, hide the leading mode / running icon on each row (e.g. assistant detail page). */
-  showSessionModeIcon?: boolean;
-  /** Main nav: list every loaded session and resolve workspace on switch (no per-workspace blocks). */
-  listAllSessions?: boolean;
-  /** Client-side filter (title, id, workspace name when listAllSessions). */
-  listFilterQuery?: string;
-}
-
-const SessionsSection: React.FC<SessionsSectionProps> = ({
-  workspaceId,
-  workspacePath,
-  remoteConnectionId = null,
-  remoteSshHost = null,
-  isActiveWorkspace: _isActiveWorkspace = true,
-  assistantLabel,
-  showSessionModeIcon = true,
-  listAllSessions = false,
-  listFilterQuery,
-}) => {
-  const { t } = useI18n('common');
-  const { setActiveWorkspace, currentWorkspace, openedWorkspacesList } = useWorkspaceContext();
-  const activeOverlay = useOverlayStore(s => s.activeOverlay);
-  const activeTabId = activeOverlay ?? AGENT_SCENE;
-  const activeBtwSessionTab = useAgentCanvasStore(state => selectActiveBtwSessionTab(state as any));
-  const activeBtwSessionData = activeBtwSessionTab?.content.data as
-    | { childSessionId: string; parentSessionId: string; workspacePath?: string }
-    | undefined;
-  const [flowChatState, setFlowChatState] = useState<FlowChatState>(() =>
-    flowChatStore.getState()
-  );
-  const [editingSessionId, setEditingSessionId] = useState<string | null>(null);
-  const [editingTitle, setEditingTitle] = useState('');
-  const [openMenuSessionId, setOpenMenuSessionId] = useState<string | null>(null);
-  const [sessionMenuPosition, setSessionMenuPosition] = useState<{ top: number; left: number } | null>(null);
-  const [runningSessionIds, setRunningSessionIds] = useState<Set<string>>(new Set());
-  const editInputRef = useRef<HTMLInputElement>(null);
-  const sessionMenuPopoverRef = useRef<HTMLDivElement>(null);
-
-  // Subscribe to state machine changes for running status
-  useEffect(() => {
-    const updateRunningSessions = () => {
-      const running = new Set<string>();
-      for (const session of flowChatState.sessions.values()) {
-        const machine = stateMachineManager.get(session.sessionId);
-        if (
-          machine &&
-          (machine.getCurrentState() === SessionExecutionState.PROCESSING ||
-            machine.getCurrentState() === SessionExecutionState.FINISHING)
-        ) {
-          running.add(session.sessionId);
-        }
-      }
-      setRunningSessionIds(running);
-    };
-
-    updateRunningSessions();
-    const unsubscribe = stateMachineManager.subscribeGlobal(() => {
-      updateRunningSessions();
-    });
-    return () => unsubscribe();
-  }, [flowChatState.sessions]);
-
-  useEffect(() => {
-    const unsub = flowChatStore.subscribe(s => setFlowChatState(s));
-    return () => unsub();
-  }, []);
-
-  useEffect(() => {
-    if (editingSessionId && editInputRef.current) {
-      editInputRef.current.focus();
-      editInputRef.current.select();
-    }
-  }, [editingSessionId]);
-
-  useEffect(() => {
-    if (!openMenuSessionId) return;
-    const handleOutside = (event: MouseEvent) => {
-      if (!sessionMenuPopoverRef.current?.contains(event.target as Node)) {
-        setOpenMenuSessionId(null);
-        setSessionMenuPosition(null);
-      }
-    };
-    document.addEventListener('mousedown', handleOutside);
-    return () => document.removeEventListener('mousedown', handleOutside);
-  }, [openMenuSessionId]);
-
-  const sessions = useMemo(
-    () =>
-      Array.from(flowChatState.sessions.values())
-        .filter((s: Session) => {
-          // Dispatcher is a standalone nav-level session — never shown in any workspace section.
-          if (s.mode === 'Dispatcher') return false;
-          if (listAllSessions) return true;
-          if (workspacePath) {
-            return sessionBelongsToWorkspaceNavRow(s, workspacePath, remoteConnectionId, remoteSshHost);
-          }
-          return !s.workspacePath;
-        })
-        .sort(compareSessionsForDisplay),
-    [flowChatState.sessions, workspacePath, remoteConnectionId, remoteSshHost, listAllSessions]
-  );
-
-  const { topLevelSessions, childrenByParent } = useMemo(() => {
-    const childMap = new Map<string, Session[]>();
-    const parents: Session[] = [];
-
-    const knownIds = new Set(sessions.map(s => s.sessionId));
-
-    for (const s of sessions) {
-      const pid = resolveSessionRelationship(s).parentSessionId;
-      if (pid && typeof pid === 'string' && pid.trim() && knownIds.has(pid)) {
-        const list = childMap.get(pid) || [];
-        list.push(s);
-        childMap.set(pid, list);
-      } else {
-        parents.push(s);
-      }
-    }
-
-    for (const [pid, list] of childMap) {
-      childMap.set(pid, [...list].sort(compareSessionsForDisplay));
-    }
-
-    return {
-      topLevelSessions: [...parents].sort(compareSessionsForDisplay),
-      childrenByParent: childMap,
-    };
-  }, [sessions]);
-
-  const visibleItems = useMemo(() => {
-    const out: Array<{ session: Session; level: 0 | 1 }> = [];
-    for (const p of topLevelSessions) {
-      out.push({ session: p, level: 0 });
-      const children = childrenByParent.get(p.sessionId) || [];
-      for (const c of children) out.push({ session: c, level: 1 });
-    }
-    return out;
-  }, [childrenByParent, topLevelSessions]);
-
-  const filteredVisibleItems = useMemo(() => {
-    const q = listFilterQuery?.trim();
-    if (!q) return visibleItems;
-    const lower = q.toLowerCase();
-    return visibleItems.filter(({ session }) => {
-      if (getTitle(session).toLowerCase().includes(lower)) return true;
-      if (session.sessionId.toLowerCase().includes(lower)) return true;
-      if (listAllSessions) {
-        const ws = findOpenedWorkspaceForSession(session, openedWorkspacesList);
-        if (ws?.name?.toLowerCase().includes(lower)) return true;
-        if (ws && getWorkspaceDisplayName(ws).toLowerCase().includes(lower)) return true;
-      }
-      return false;
-    });
-  }, [visibleItems, listFilterQuery, listAllSessions, openedWorkspacesList]);
-
-  const activeSessionId = flowChatState.activeSessionId;
-
-  const handleSwitch = useCallback(
-    async (sessionId: string) => {
-      if (editingSessionId) return;
-      try {
-        const session = flowChatStore.getState().sessions.get(sessionId);
-        const relationship = resolveSessionRelationship(session);
-        const parentSessionId = relationship.parentSessionId;
-        const resolvedWorkspaceId =
-          listAllSessions && session
-            ? findOpenedWorkspaceForSession(session, openedWorkspacesList)?.id
-            : workspaceId;
-        const mustActivateWorkspace =
-          Boolean(resolvedWorkspaceId) && resolvedWorkspaceId !== currentWorkspace?.id;
-        const activateWorkspace = mustActivateWorkspace
-          ? async (targetWorkspaceId: string) => {
-              await setActiveWorkspace(targetWorkspaceId);
-            }
-          : undefined;
-
-        if (relationship.canOpenInAuxPane && parentSessionId && session) {
-          await openMainSession(parentSessionId, {
-            workspaceId: resolvedWorkspaceId,
-            activateWorkspace,
-          });
-          openBtwSessionInAuxPane({
-            childSessionId: sessionId,
-            parentSessionId,
-            workspacePath: session.workspacePath,
-          });
-          return;
-        }
-
-        if (sessionId === activeSessionId) {
-          await openMainSession(sessionId, {
-            workspaceId: resolvedWorkspaceId,
-            activateWorkspace,
-          });
-          return;
-        }
-
-        await openMainSession(sessionId, {
-          workspaceId: resolvedWorkspaceId,
-          activateWorkspace,
-        });
-        window.dispatchEvent(
-          new CustomEvent('flowchat:switch-session', { detail: { sessionId } })
-        );
-      } catch (err) {
-        log.error('Failed to switch session', err);
-      }
-    },
-    [
-      activeSessionId,
-      editingSessionId,
-      setActiveWorkspace,
-      workspaceId,
-      currentWorkspace?.id,
-      listAllSessions,
-      openedWorkspacesList,
-    ]
-  );
-
-  const resolveSessionTitle = useCallback(
-    (session: Session): string => {
-      const rawTitle = getTitle(session);
-      const matched = rawTitle.match(/^(?:新建会话|New Session)\s*(\d+)$/i);
-      if (!matched) return rawTitle;
-
-      const mode = resolveSessionModeType(session);
-      const label =
-        mode === 'cowork'
-          ? t('nav.sessions.newCoworkSession')
-          : mode === 'claw'
-            ? t('nav.sessions.newClawSession')
-            : t('nav.sessions.newCodeSession');
-      return `${label} ${matched[1]}`;
-    },
-    [t]
-  );
-
-  const handleMenuOpen = useCallback(
-    (e: React.MouseEvent, sessionId: string) => {
-      e.stopPropagation();
-      if (openMenuSessionId === sessionId) {
-        setOpenMenuSessionId(null);
-        setSessionMenuPosition(null);
-        return;
-      }
-      const btn = e.currentTarget as HTMLElement;
-      const rect = btn.getBoundingClientRect();
-      const viewportPadding = 8;
-      const estimatedWidth = 160;
-      const maxLeft = window.innerWidth - estimatedWidth - viewportPadding;
-      setSessionMenuPosition({
-        top: Math.max(viewportPadding, rect.bottom + 4),
-        left: Math.max(viewportPadding, Math.min(rect.left, maxLeft)),
-      });
-      setOpenMenuSessionId(sessionId);
-    },
-    [openMenuSessionId]
-  );
-
-  const handleDelete = useCallback(
-    async (e: React.MouseEvent, sessionId: string) => {
-      e.stopPropagation();
-      try {
-        await flowChatManager.deleteChatSession(sessionId);
-      } catch (err) {
-        log.error('Failed to delete session', err);
-      }
-    },
-    []
-  );
-
-  const handleStartEdit = useCallback(
-    (e: React.MouseEvent, session: Session) => {
-      e.stopPropagation();
-      setEditingSessionId(session.sessionId);
-      setEditingTitle(resolveSessionTitle(session));
-    },
-    [resolveSessionTitle]
-  );
-
-  const handleConfirmEdit = useCallback(async () => {
-    if (!editingSessionId) return;
-    const trimmed = editingTitle.trim();
-    if (trimmed) {
-      try {
-        await flowChatManager.renameChatSessionTitle(editingSessionId, trimmed);
-      } catch (err) {
-        log.error('Failed to update session title', err);
-      }
-    }
-    setEditingSessionId(null);
-    setEditingTitle('');
-  }, [editingSessionId, editingTitle]);
-
-  const handleCancelEdit = useCallback(() => {
-    setEditingSessionId(null);
-    setEditingTitle('');
-  }, []);
-
-  const handleEditKeyDown = useCallback(
-    (e: React.KeyboardEvent) => {
-      if (e.key === 'Enter') {
-        e.preventDefault();
-        handleConfirmEdit();
-      } else if (e.key === 'Escape') {
-        e.preventDefault();
-        handleCancelEdit();
-      }
-    },
-    [handleConfirmEdit, handleCancelEdit]
-  );
-
-  if (topLevelSessions.length === 0) {
-    return null;
-  }
-
-  if (filteredVisibleItems.length === 0 && listFilterQuery?.trim()) {
-    return (
-      <div className="bitfun-nav-panel__inline-list bitfun-nav-panel__inline-list--filter-empty">
-        <div className="bitfun-nav-panel__filter-empty">{t('nav.sessionCapsule.filterNoMatch')}</div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="bitfun-nav-panel__inline-list">
-      {filteredVisibleItems.map(({ session, level }) => {
-          const isEditing = editingSessionId === session.sessionId;
-          const relationship = resolveSessionRelationship(session);
-          const isBtwChild = level === 1 && relationship.isBtw;
-          const sessionModeKey = resolveSessionModeType(session);
-          const sessionTitle = resolveSessionTitle(session);
-          const parentSessionId = relationship.parentSessionId;
-          const parentSession = parentSessionId ? flowChatState.sessions.get(parentSessionId) : undefined;
-          const parentTitle = parentSession ? resolveSessionTitle(parentSession) : '';
-          const parentTurnIndex = relationship.origin?.parentTurnIndex;
-          const rowWorkspace = listAllSessions
-            ? findOpenedWorkspaceForSession(session, openedWorkspacesList)
-            : undefined;
-          const contextLabel = listAllSessions
-            ? (rowWorkspace ? getWorkspaceDisplayName(rowWorkspace) : '')
-            : (assistantLabel?.trim() ?? '');
-          const showContextInTooltip = contextLabel.length > 0;
-          const showRichTooltip = showContextInTooltip || isBtwChild;
-          const tooltipContent = showRichTooltip ? (
-            <div className="bitfun-nav-panel__inline-item-tooltip">
-              <div className="bitfun-nav-panel__inline-item-tooltip-title">{sessionTitle}</div>
-              {showContextInTooltip ? (
-                <div className="bitfun-nav-panel__inline-item-tooltip-meta">
-                  {listAllSessions
-                    ? t('nav.sessions.sessionContext', { name: contextLabel })
-                    : t('nav.sessions.assistantOwner', { name: contextLabel })}
-                </div>
-              ) : null}
-              {isBtwChild ? (
-                <div className="bitfun-nav-panel__inline-item-tooltip-meta">
-                  {`来自 ${parentTitle || '父会话'}${parentTurnIndex ? ` · 第 ${parentTurnIndex} 轮` : ''}`}
-                </div>
-              ) : null}
-            </div>
-          ) : (
-            sessionTitle
-          );
-          const SessionIcon =
-            sessionModeKey === 'cowork'
-              ? ListTodo
-              : sessionModeKey === 'claw'
-                ? Sparkles
-                : Code2;
-          const isRunning = runningSessionIds.has(session.sessionId);
-          const isRowActive = activeBtwSessionData?.childSessionId
-            ? session.sessionId === activeBtwSessionData.childSessionId
-            : activeTabId === AGENT_SCENE && session.sessionId === activeSessionId;
-          const row = (
-            <div
-              className={[
-                'bitfun-nav-panel__inline-item',
-                level === 1 && 'is-child',
-                isBtwChild && 'is-btw-child',
-                isRowActive && 'is-active',
-                isEditing && 'is-editing',
-                openMenuSessionId === session.sessionId && 'is-menu-open',
-              ]
-                .filter(Boolean)
-                .join(' ')}
-              onClick={() => handleSwitch(session.sessionId)}
-            >
-              {showSessionModeIcon && !isBtwChild ? (
-                isRunning ? (
-                  <Loader2
-                    size={14}
-                    className={[
-                      'bitfun-nav-panel__inline-item-icon',
-                      'is-running',
-                    ].join(' ')}
-                  />
-                ) : (
-                  <SessionIcon
-                    size={14}
-                    className={[
-                      'bitfun-nav-panel__inline-item-icon',
-                      sessionModeKey === 'cowork'
-                        ? 'is-cowork'
-                        : sessionModeKey === 'claw'
-                          ? 'is-claw'
-                          : 'is-code',
-                    ].join(' ')}
-                  />
-                )
-              ) : null}
-
-              {isEditing ? (
-                <div className="bitfun-nav-panel__inline-item-edit" onClick={e => e.stopPropagation()}>
-                  <Input
-                    ref={editInputRef}
-                    className="bitfun-nav-panel__inline-item-edit-field"
-                    variant="default"
-                    inputSize="small"
-                    value={editingTitle}
-                    onChange={e => setEditingTitle(e.target.value)}
-                    onKeyDown={handleEditKeyDown}
-                    onBlur={handleConfirmEdit}
-                  />
-                  <IconButton
-                    variant="success"
-                    size="xs"
-                    className="bitfun-nav-panel__inline-item-edit-btn confirm"
-                    onClick={e => { e.stopPropagation(); handleConfirmEdit(); }}
-                    tooltip={t('nav.sessions.confirmEdit')}
-                    tooltipPlacement="top"
-                  >
-                    <Check size={11} />
-                  </IconButton>
-                  <IconButton
-                    variant="default"
-                    size="xs"
-                    className="bitfun-nav-panel__inline-item-edit-btn cancel"
-                    onMouseDown={e => { e.preventDefault(); e.stopPropagation(); handleCancelEdit(); }}
-                    tooltip={t('nav.sessions.cancelEdit')}
-                    tooltipPlacement="top"
-                  >
-                    <X size={11} />
-                  </IconButton>
-                </div>
-              ) : (
-                <>
-                  <span className="bitfun-nav-panel__inline-item-main">
-                    <span className="bitfun-nav-panel__inline-item-label">{sessionTitle}</span>
-                    {isBtwChild ? (
-                      <span className="bitfun-nav-panel__inline-item-btw-badge">btw</span>
-                    ) : null}
-                  </span>
-                  <div className="bitfun-nav-panel__inline-item-actions">
-                    <button
-                      type="button"
-                      className={`bitfun-nav-panel__inline-item-action-btn${openMenuSessionId === session.sessionId ? ' is-open' : ''}`}
-                      onClick={e => handleMenuOpen(e, session.sessionId)}
-                    >
-                      <MoreHorizontal size={12} />
-                    </button>
-                  </div>
-                  {openMenuSessionId === session.sessionId && sessionMenuPosition && createPortal(
-                    <div
-                      ref={sessionMenuPopoverRef}
-                      className="bitfun-nav-panel__inline-item-menu-popover"
-                      role="menu"
-                      data-bitfun-ignore-session-capsule-outside
-                      style={{ top: `${sessionMenuPosition.top}px`, left: `${sessionMenuPosition.left}px` }}
-                    >
-                      <button
-                        type="button"
-                        className="bitfun-nav-panel__inline-item-menu-item"
-                        onClick={e => { setOpenMenuSessionId(null); handleStartEdit(e, session); }}
-                      >
-                        <Pencil size={13} />
-                        <span>{t('nav.sessions.rename')}</span>
-                      </button>
-                      <button
-                        type="button"
-                        className="bitfun-nav-panel__inline-item-menu-item is-danger"
-                        onClick={e => { setOpenMenuSessionId(null); void handleDelete(e, session.sessionId); }}
-                      >
-                        <Trash2 size={13} />
-                        <span>{t('nav.sessions.delete')}</span>
-                      </button>
-                    </div>,
-                    document.body
-                  )}
-                </>
-              )}
-            </div>
-          );
-          return isEditing || openMenuSessionId !== null ? row : (
-            <Tooltip key={session.sessionId} content={tooltipContent} placement="right" followCursor>
-              {row}
-            </Tooltip>
-          );
-        })}
-    </div>
-  );
-};
-
-export default SessionsSection;
