+++ src/web-ui/src/app/components/index.ts
@@ -2,7 +2,6 @@
  * App component exports.
  */
 
-export { default as NavPanel } from './NavPanel/NavPanel';
 export type { SceneTabId, OverlaySceneId, SceneId } from '../overlay/types';
 export * from './GalleryLayout';
 export * from './panels';
