+++ src/web-ui/src/app/hooks/index.ts
@@ -6,7 +6,6 @@ export * from './useApp';
 export * from './useWindowControls';
 export * from './useSceneManager';
 export * from './useOverlayManager';
-export * from './useNavHistory';
 export * from './useCurrentSessionTitle';
 export * from './useCurrentSettingsTabTitle';
 export * from './useGlobalSceneShortcuts';
