--- src/web-ui/src/flow_chat/components/ImageAnalysisCard.scss
@@ -1,269 +1,0 @@
-/**
- * Image analysis card styles
- */
-
-.image-analysis-card {
-  background: var(--color-surface);
-  border: 1px solid var(--color-border);
-  border-radius: 8px;
-  padding: 12px;
-  margin: 8px 0;
-  transition: all 0.2s ease;
-  
-  &[data-status="analyzing"] {
-    border-color: var(--color-primary);
-    background: linear-gradient(
-      90deg, 
-      var(--color-surface) 0%, 
-      rgba(var(--color-primary-rgb), 0.05) 50%, 
-      var(--color-surface) 100%
-    );
-    background-size: 200% 100%;
-    animation: shimmer 2s infinite;
-  }
-  
-  &[data-status="completed"] {
-    border-color: var(--color-success);
-  }
-  
-  &[data-status="error"] {
-    border-color: var(--color-error);
-  }
-}
-
-@keyframes shimmer {
-  0% {
-    background-position: -200% 0;
-  }
-  100% {
-    background-position: 200% 0;
-  }
-}
-
-.image-analysis-card__header {
-  display: flex;
-  align-items: flex-start;
-  gap: 12px;
-}
-
-.image-analysis-card__thumbnail {
-  width: 60px;
-  height: 60px;
-  flex-shrink: 0;
-  border-radius: 6px;
-  overflow: hidden;
-  background: var(--color-background);
-  border: 1px solid var(--color-border);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  
-  img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }
-}
-
-.image-analysis-card__thumbnail-placeholder {
-  color: var(--color-text-secondary);
-}
-
-.image-analysis-card__info {
-  flex: 1;
-  min-width: 0;
-}
-
-.image-analysis-card__filename {
-  font-size: 14px;
-  font-weight: 500;
-  color: var(--color-text);
-  margin-bottom: 6px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-// Status variants
-.image-analysis-card__status {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  font-size: 12px;
-  
-  &.analyzing {
-    color: var(--color-primary);
-    
-    .spinner {
-      animation: spin 1s linear infinite;
-    }
-  }
-  
-  &.completed {
-    color: var(--color-success);
-    
-    .time {
-      margin-left: 4px;
-      color: var(--color-text-secondary);
-    }
-  }
-  
-  &.error {
-    color: var(--color-error);
-    
-    .retry-btn {
-      margin-left: 8px;
-      padding: 2px 8px;
-      font-size: 11px;
-      background: var(--color-error);
-      color: white;
-      border: none;
-      border-radius: 4px;
-      cursor: pointer;
-      transition: opacity 0.2s;
-      
-      &:hover {
-        opacity: 0.8;
-      }
-    }
-  }
-}
-
-@keyframes spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.image-analysis-card__content {
-  margin-top: 12px;
-  padding-top: 12px;
-  border-top: 1px solid var(--color-border);
-}
-
-.image-analysis-card__summary {
-  display: flex;
-  align-items: flex-start;
-  gap: 8px;
-  font-size: 13px;
-  line-height: 1.5;
-  color: var(--color-text);
-  
-  .summary-icon {
-    flex-shrink: 0;
-    margin-top: 2px;
-    color: var(--color-primary);
-  }
-}
-
-.image-analysis-card__expand-btn {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  margin-top: 8px;
-  padding: 4px 8px;
-  font-size: 12px;
-  color: var(--color-primary);
-  background: transparent;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  transition: background 0.2s;
-  
-  &:hover {
-    background: rgba(var(--color-primary-rgb), 0.1);
-  }
-}
-
-// Detailed content
-.image-analysis-card__detailed {
-  margin-top: 12px;
-  padding: 12px;
-  background: var(--color-background);
-  border-radius: 6px;
-  
-  .detail-section {
-    margin-bottom: 12px;
-    
-    &:last-child {
-      margin-bottom: 0;
-    }
-    
-    h4 {
-      font-size: 12px;
-      font-weight: 600;
-      color: var(--color-text-secondary);
-      margin: 0 0 6px 0;
-      text-transform: uppercase;
-      letter-spacing: 0.5px;
-    }
-    
-    p {
-      font-size: 13px;
-      line-height: 1.6;
-      color: var(--color-text);
-      margin: 0;
-    }
-  }
-  
-  .tags {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 6px;
-    
-    .tag {
-      display: inline-block;
-      padding: 4px 10px;
-      font-size: 12px;
-      background: var(--color-surface);
-      color: var(--color-text);
-      border: 1px solid var(--color-border);
-      border-radius: 12px;
-    }
-  }
-  
-  .metadata {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    font-size: 12px;
-    color: var(--color-text-secondary);
-    
-    .meta-separator {
-      opacity: 0.5;
-    }
-  }
-}
-
-// Error message
-.image-analysis-card__error {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  margin-top: 12px;
-  padding: 8px 12px;
-  background: rgba(var(--color-error-rgb), 0.1);
-  border-radius: 6px;
-  font-size: 12px;
-  color: var(--color-error);
-}
-
-// Dark mode support
-@media (prefers-color-scheme: dark) {
-  .image-analysis-card {
-    --color-surface: #1e1e1e;
-    --color-background: #121212;
-    --color-border: #333;
-    --color-text: #e0e0e0;
-    --color-text-secondary: #a0a0a0;
-    --color-primary: #60a5fa;
-    --color-primary-rgb: 96, 165, 250;
-    --color-success: #22c55e;
-    --color-error: #ef4444;
-    --color-error-rgb: 239, 68, 68;
-  }
-}
-
