--- src/web-ui/src/infrastructure/services/api/index.ts
@@ -1,5 +1,0 @@
- 
-
-export { default as aiService } from './aiService';
-export { default as contextService } from './contextService';
-
