+++ src/web-ui/src/shared/notification-system/styles/_match-nav-search-surface.scss
@@ -1,5 +1,5 @@
 /**
- * Surfaces aligned with NavSearchDialog card (scene background, layered shadow, radius).
+ * Surfaces aligned with GlobalSearchDialog card (scene background, layered shadow, radius).
  */
 @use '../../../component-library/styles/tokens.scss' as *;
