+++ src/web-ui/src/shared/utils/tabUtils.ts
@@ -7,6 +7,7 @@ import { enqueuePendingTab } from '@/shared/services/pendingTabQueue';
 import { resolveAndFocusOpenTarget } from '@/shared/services/sceneOpenTargetResolver';
 import type { OpenSource } from '@/shared/services/sceneOpenTargetResolver';
 import { TAB_EVENTS } from '@/app/components/panels/content-canvas/types';
+import { useOverlayStore } from '@/app/stores/overlayStore';
 export type TabTargetMode = 'agent' | 'project';
 
 export interface TabCreationOptions {
@@ -202,8 +203,7 @@ export function createConfigCenterTab(
   _initialTab: 'models' | 'agents' = 'models',
   _mode: 'agent' | 'project' = 'agent'
 ): void {
-  // Settings is now an independent scene — open via event bus.
-  window.dispatchEvent(new CustomEvent('scene:open', { detail: { sceneId: 'settings' } }));
+  useOverlayStore.getState().openOverlay('settings');
 }
 
 export function createTerminalTab(
